[ISSUE #1906] The BooleanConstantExpression might lead to class loading deadlock
[ISSUE #1906] The BooleanConstantExpression might lead to class loading deadlock
